requirements.txt
langchain-openai>=0.0.6
python-dotenv>=1.0.0
python-docx>=0.8.11
pdfminer.six>=20221105